+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - :sparkles: `HttpRequestHandler#clearStaticResCache` 方法改为 default 方法。
 - :sparkles: 支持空 `websocket` 包
 - :sparkles: 完善 `HexUtils` 工具类。
+- :sparkles: 添加 JsonAdapter，支持 `Jackson`, `Fastjson1`, `Fastjson2`, `Gson`, `hutool-json`。
 - :sparkles: copy 并优化 `TimedCache`，用于处理 mqtt 保留 session。
 - :arrow_up: 依赖升级